qwen-vl/requirements.txt
Pillow
requests
tqdm
transformers==4.36.2
tiktoken
matplotlib
transformers_stream_generator
einops
torchvision
safetensors
accelerate==0.25.0